move test helpers into board conditions class
